Changing some tests to other test suite. Fixing when alter table adds a new constraint related to foreign key.
