+++ tests/sqlite/schema/test_sqlite_alter_table.py
@@ -2,9 +2,13 @@ from src.masoniteorm.schema.Blueprint import Blueprint
 from src.masoniteorm.schema.grammars import GrammarFactory
 from src.masoniteorm.schema import Schema
 import unittest
+import inspect
 
 
 class TestSqliteAlterGrammar(unittest.TestCase):
+
+    maxDiff = None
+
     def setUp(self):
         self.schema = Schema.dry().on("sqlite")
 
@@ -51,30 +55,54 @@ class TestSqliteAlterGrammar(unittest.TestCase):
 
         sql = """ALTER TABLE "users" ADD CONSTRAINT name_unique UNIQUE("name")"""
         self.assertEqual(blueprint.to_sql(), sql)
+    
+    def test_foreign_key_constraint(self):
+        with self.schema.create("users") as blueprint:
+            blueprint.unsigned("user_id")
+            blueprint.foreign("user_id").references("id").on("profile")
+            blueprint.foreign("fruit_id").references("id").on("fruit")
+
+        sql = getattr(
+            self, inspect.currentframe().f_code.co_name.replace("test_", "")
+        )()
+        self.assertEqual(blueprint.to_sql(), sql)
+    
+    def test_a_second_foreign_key_constraint(self):
+        with self.schema.table('videos') as blueprint:
+            blueprint.unsigned_integer('playlist_id').nullable()
+            blueprint.foreign('playlist_id').references('id').on('playlists')
+
+        sql = getattr(
+            self, inspect.currentframe().f_code.co_name.replace("test_", "")
+        )()
+        self.assertEqual(blueprint.to_sql(), sql)
 
     def test_can_alter_foreign_key(self):
         with self.schema.table("users") as blueprint:
+            blueprint.unsigned_integer('profile_id').nullable()
             blueprint.foreign("profile_id").references("id").on("profile")
 
-        sql = """ALTER TABLE "users" ADD CONSTRAINT users_profile_id_foreign FOREIGN KEY ("profile_id") REFERENCES "profile"("id")"""
+        sql = """ALTER TABLE "users" ADD "profile_id" UNSIGNED INT, CONSTRAINT users_profile_id_foreign REFERENCES "profile"("id")"""
         self.assertEqual(blueprint.to_sql(), sql)
 
     def test_can_alter_foreign_key_with_on_delete(self):
         with self.schema.table("users") as blueprint:
+            blueprint.unsigned_integer('profile_id').nullable()
             blueprint.foreign("profile_id").references("id").on("profile").on_delete(
                 "cascade"
             )
 
-        sql = """ALTER TABLE "users" ADD CONSTRAINT users_profile_id_foreign FOREIGN KEY ("profile_id") REFERENCES "profile"("id") ON DELETE CASCADE"""
+        sql = """ALTER TABLE "users" ADD "profile_id" UNSIGNED INT, CONSTRAINT users_profile_id_foreign REFERENCES "profile"("id") ON DELETE CASCADE"""
         self.assertEqual(blueprint.to_sql(), sql)
 
     def test_can_alter_foreign_key_with_on_update(self):
         with self.schema.table("users") as blueprint:
+            blueprint.unsigned_integer('profile_id').nullable()
             blueprint.foreign("profile_id").references("id").on("profile").on_update(
                 "cascade"
             )
 
-        sql = """ALTER TABLE "users" ADD CONSTRAINT users_profile_id_foreign FOREIGN KEY ("profile_id") REFERENCES "profile"("id") ON UPDATE CASCADE"""
+        sql = """ALTER TABLE "users" ADD "profile_id" UNSIGNED INT, CONSTRAINT users_profile_id_foreign REFERENCES "profile"("id") ON UPDATE CASCADE"""
         self.assertEqual(blueprint.to_sql(), sql)
 
     def test_can_alter_modify_column(self):
@@ -83,3 +111,33 @@ class TestSqliteAlterGrammar(unittest.TestCase):
 
         sql = """ALTER TABLE "users" DROP COLUMN "name\""""
         self.assertEqual(blueprint.to_sql(), sql)
+
+    def foreign_key_constraint(self):
+        """
+        with self.schema.create("users") as blueprint:
+            blueprint.integer("user_id").unsigned()
+            blueprint.foreign('user_id').references('id').on('profile')
+            blueprint.foreign('fruit_id').references('id').on('fruit')
+        """
+
+        return (
+            """CREATE TABLE "users" ("""
+            """\"user_id" UNSIGNED INT NOT NULL, """
+            """ADD CONSTRAINT users_user_id_foreign FOREIGN KEY ("user_id") REFERENCES "profile"("id"), """
+            """ADD CONSTRAINT users_fruit_id_foreign FOREIGN KEY ("fruit_id") REFERENCES "fruit"("id")"""
+            """)"""
+        )
+    
+    def a_second_foreign_key_constraint(self):
+        """
+        with self.schema.table('videos') as table:
+            table.unsigned_integer('playlist_id').nullable()
+            table.foreign('playlist_id').references('id').on('playlists')
+        """
+
+        return (
+            """ALTER TABLE "videos" """
+            """ADD "playlist_id" UNSIGNED INT, """
+            """CONSTRAINT videos_playlist_id_foreign """
+            """REFERENCES "playlists"("id")"""
+        )
